Added class for executing tasks
